requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.21.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
apscheduler==3.11.2
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.5
blinker==1.9.0
build==1.4.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.1.8
coloredlogs==15.0.1
crewai==1.9.3
crewai-tools==1.9.3
cryptography==46.0.4
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring-parser==0.17.0
durationpy==0.10
email-validator==2.3.0
et-xmlfile==2.0.0
fastapi==0.128.0
fastapi-sso==0.20.0
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
gitpython==3.1.46
googleapis-common-protos==1.72.0
greenlet==3.3.1
groq==0.37.1
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.1
humanfriendly==10.0
identify==2.6.16
idna==3.11
importlib-metadata==8.7.1
importlib-resources==6.5.2
instructor==1.12.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.3
json-repair==0.25.3
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
lancedb==0.5.7
langchain-chroma==1.0.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.8
langchain-groq==1.1.2
langchain-huggingface==1.2.0
langchain-milvus==0.3.3
langchain-text-splitters==1.1.0
langsmith==0.6.8
litellm==1.75.3
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.2
mcp==1.23.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
mypy-extensions==1.1.0
narwhals==2.16.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.83.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pdfminer-six==20251230
pdfplumber==0.11.9
pillow==12.1.0
platformdirs==4.5.1
portalocker==2.7.0
posthog==5.4.0
pre-commit==4.5.1
propcache==0.4.1
protobuf==5.29.5
py==1.11.0
pyarrow==23.0.0
pybase64==1.4.3
pycparser==3.0
pydantic==2.11.10
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydeck==0.9.1
pygments==2.19.2
pyjwt==2.11.0
pylance==0.9.18
pymilvus==2.6.8
pymupdf==1.26.7
pypdf==6.6.2
pypdfium2==5.3.0
pypika==0.51.0
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.22
pytube==15.0.0
pytz==2025.2
pyyaml==6.0.3
-e file:///home/beamx/Desktop/rag_crew
ratelimiter==1.2.0.post0
referencing==0.37.0
regex==2024.9.11
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==14.3.2
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
semver==3.0.4
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
sqlalchemy==2.0.46
sse-starlette==3.2.0
starlette==0.50.0
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.2
tomli-w==1.1.0
torch==2.10.0
tornado==6.5.4
tqdm==4.67.2
transformers==4.46.3
triton==3.6.0
typer==0.21.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid-utils==0.14.0
uv==0.9.28
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.36.1
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
yarl==1.22.0
youtube-transcript-api==1.2.4
zipp==3.23.0
zstandard==0.25.0
